# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Freeable Domains API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project settings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{id}/summary:
    get:
      operationId: getProjectSummary
      tags: [projects]
      summary: Get project summary with recent activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSummary"

  # ── Deployments ───────────────────────────────────────────────────────────
...
  /projects/{id}/env:
    get:
      operationId: listEnvVars
      tags: [env]
      summary: List environment variables for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Env vars
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EnvVar"
    post:
      operationId: createEnvVar
      tags: [env]
      summary: Create an environment variable
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnvVarInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvVar"
  /env/{id}:
    put:
      operationId: updateEnvVar
      tags: [env]
      summary: Update an environment variable
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnvVarUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvVar"
    delete:
      operationId: deleteEnvVar
      tags: [env]
      summary: Delete an environment variable
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── GitHub Import ─────────────────────────────────────────────────────────
  /github/detect:
    post:
      operationId: detectFramework
      tags: [github]
      summary: Detect framework from a GitHub repo URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubDetectInput"
      responses:
        "200":
          description: Detection result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FrameworkDetectionResult"
  /github/import:
    post:
      operationId: importFromGithub
      tags: [github]
      summary: Import a project from GitHub and start deployment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubImportInput"
      responses:
        "201":
          description: Project created and deployment triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GithubImportResult"

  # ── Infrastructure ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        repoUrl:
          type: ["string", "null"]
        branch:
          type: string
        framework:
          type: ["string", "null"]
        packageManager:
          type: ["string", "null"]
        buildCommand:
          type: ["string", "null"]
        outputDir:
          type: ["string", "null"]
        status:
          type: string
          enum: [idle, building, deploying, active, failed, stopped]
        deploymentUrl:
          type: ["string", "null"]
        sessionId:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - branch
        - status
        - createdAt
        - updatedAt
    ProjectInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        repoUrl:
          type: string
        branch:
          type: string
        framework:
          type: string
        buildCommand:
          type: string
        outputDir:
          type: string
        sessionId:
          type: string
      required: [name]
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        branch:
          type: string
        buildCommand:
          type: string
        outputDir:
          type: string
        status:
          type: string
    ProjectSummary:
      type: object
      properties:
        project:
          $ref: "#/components/schemas/Project"
        latestDeployment:
          $ref: "#/components/schemas/Deployment"
        deploymentCount:
          type: integer
        successfulDeployments:
          type: integer
        failedDeployments:
          type: integer
        avgBuildTimeSeconds:
          type: number
        domainCount:
          type: integer
        envVarCount:
          type: integer
        requestsToday:
          type: integer
        bandwidthTodayGb:
          type: number
      required:
        - project
        - deploymentCount
        - successfulDeployments
        - failedDeployments
        - avgBuildTimeSeconds
        - domainCount
        - envVarCount
        - requestsToday
        - bandwidthTodayGb

    # ── Deployments ─────────────────────────────────────────────────────────
    Deployment:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        projectName:
          type: ["string", "null"]
        status:
          type: string
          enum: [queued, cloning, installing, building, deploying, ready, failed, cancelled, rolled_back]
        url:
          type: ["string", "null"]
        commitSha:
          type: ["string", "null"]
        commitMessage:
          type: ["string", "null"]
        branch:
          type: string
        buildDurationSeconds:
          type: ["number", "null"]
        triggeredBy:
          type: string
          enum: [manual, github_push, api, rollback]
        isProduction:
          type: boolean
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
      required:
        - id
        - projectId
        - status
        - branch
        - triggeredBy
        - isProduction
        - createdAt
...
    EnvVar:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        key:
          type: string
        value:
          type: string
        encrypted:
          type: boolean
        scope:
          type: string
          enum: [build, runtime, all]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, projectId, key, value, encrypted, scope, createdAt, updatedAt]
    EnvVarInput:
      type: object
      properties:
        key:
          type: string
          minLength: 1
        value:
          type: string
        encrypted:
          type: boolean
        scope:
          type: string
          enum: [build, runtime, all]
      required: [key, value]
    EnvVarUpdate:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
        encrypted:
          type: boolean
        scope:
          type: string

    # ── GitHub ──────────────────────────────────────────────────────────────
    GithubDetectInput:
      type: object
      properties:
        repoUrl:
          type: string
          minLength: 1
      required: [repoUrl]
    FrameworkDetectionResult:
      type: object
      properties:
        repoUrl:
          type: string
        framework:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
        packageManager:
          type: ["string", "null"]
        buildCommand:
          type: ["string", "null"]
        outputDir:
          type: ["string", "null"]
        detectedFiles:
          type: array
          items:
            type: string
      required: [repoUrl, detectedFiles]
    GithubImportInput:
      type: object
      properties:
        repoUrl:
          type: string
          minLength: 1
        name:
          type: string
        branch:
          type: string
        sessionId:
          type: string
      required: [repoUrl]
    GithubImportResult:
      type: object
      properties:
        project:
          $ref: "#/components/schemas/Project"
        deployment:
          $ref: "#/components/schemas/Deployment"
        detectedFramework:
          type: ["string", "null"]
      required: [project, deployment]

    # ── Infrastructure ──────────────────────────────────────────────────────
...